+++ CHANGES.md
@@ -6,6 +6,8 @@
 - [FIXED] Ensure IAM API key can be correctly changed.
 - [FIXED] Callback with an error when a user cannot be authenticated using IAM.
 - [FIXED] Ensure authorization tokens are not unnecessarily requested.
+- [IMPROVED] Do not apply cookie authentication by default in the case that no
+  credentials are provided.
 - [IMPROVED] Preemptively renew authorization tokens that are due to expire.
 
 # 4.2.1 (2019-08-29)